+++ src/main/java/com/ontologycentral/osmwrap/ApiConstants.java
@@ -14,6 +14,7 @@ public final class ApiConstants {
     public static final int SEARCH_READ_TIMEOUT = 30000;
     public static final int POI_READ_TIMEOUT = 30000;
     public static final int GEOMETRY_READ_TIMEOUT = 30000;
+    public static final int AROUND_READ_TIMEOUT = 30000;
 
     private ApiConstants() {
         // Utility class